pkg/logutil: move "pkg/logger" to "logutil"
