+++ pkg/logutil/doc.go
@@ -12,5 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package logger implements various logging utilities.
-package logger
+// Package logutil includes utilities to facilitate logging.
+package logutil
