+++ pkg/logutil/package_logger.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package logger
+package logutil
 
 import (
 	"github.com/coreos/pkg/capnslog"
